+++ remix/app/api/utils/things/things.ts
@@ -2862,6 +2862,7 @@ export type AddCommentInput =
       images?: unknown;
       listing?: unknown;
       thing?: unknown;
+			mediaLayout?: unknown;
       tags?: unknown;
 			shareId?: unknown;
 	  };